org-download.el (org-download-clipboard): Correct ImageMagick's executable name on Windows
On default installations of ImageMagick on Windows, executables such as `convert.exe` are considered a legacy feature (perhaps due to name collisions with system programs) and as such are not included unless the user asks for it during the installation. Instead, these utilities are treated as subcommands that can be accessed through `magick.exe`.